refactor: put movies data and interface into separate files
