+++ react-components/src/components/MovieCard/MovieCard.tsx
@@ -1,17 +1,6 @@
 import React from 'react';
+import { IMovie } from 'types';
 import './MovieCard.css';
-
-export interface IMovie {
-  backdrop_path: string;
-  genres: Array<string>;
-  id: number;
-  original_language: string;
-  overview: string;
-  release_date: string;
-  runtime: number;
-  title: string;
-  vote_average: number;
-}
 
 interface MovieCardProps {
   movie: IMovie;
