+++ react-components/src/types.ts
@@ -1,0 +1,11 @@
+export interface IMovie {
+  backdrop_path: string;
+  genres: Array<string>;
+  id: number;
+  original_language: string;
+  overview: string;
+  release_date: string;
+  runtime: number;
+  title: string;
+  vote_average: number;
+}
